VAN-2096 fix jq setup
On aws pipelines, jq seems already installed but it was jq 1.5
we were upgrading it to 1.6 (for consistency)
but the PATH variable was set in a way that the newly installed one
was getting overriden. So we fix it by overriding the previous one.

Also the format `jq -aRs` is not backward compatible so changing it to
`jq -aRs .` since that makes it more resilient and consistent
to the other occurrences