VAN-2748 Copy TF state instead of using push
Terraform state must be present in the working directory
before running init. It must be called terraform.tfstate.

We used terraform push command to achieve this. This
command is primarily for working with remote backends.
It performs additional checks.

Unfortunatelly it fails on TF >= 1.2.0
It requires init to be done first.
This is not compatible with our use to initialize the working
directory before init.

Since we do not really require the push as we managed the state
ourselves we can instead just copy the state file to the working directoy
to achieve the same result.