VAN-3245 Refactor pipeline modules for re-use
- Added an assets/scripts directory that contains some common
functions as bash scripts that can be re-used across services
and clouds.
- Added a common directory that contains pipeline modules that
can be re-used across a services within one cloud type.
- In order to facilitate smoother local development and db
seeding, the common modules will be symlinked into the appropriate
pipeline module directories with the prefix "cmn-"
- For GCP, add a custom container - cpi-config - that is used
at various points during the config. This container has some
basic linux tools in it (including JQ) that are used in the
config steps. It also has the path set for the assets/scripts
dir so those scripts can be called without full path.
- Added a make target - cloud-builders - which will build and
push all the containers in the cloud-builders directory up to
dockerhub. Can be called with an option "TAG" variable which
will then tag all the containers. By default containers are
tagged with the commit SHA.