VAN-3243 Add pipeline module to call make targets
Added "make-target" pipeline module for AWS, Azure and GCP. This
module allows the integration pipeline to call 1 or more targets
in a make file in sequence.

In order to have the secret env vars available to any downstream
processes (like references in the Makefile), needed to add an
"env" stanza to any steps/tasks that needed them. Also did this
for the env-files module as it won't actually work without this.

For GCP, this pipeline module also includes a new cloud-builder
container.

Also, recategorized "bash" to "utility" as I think it shows up
better in the UI this way.