VAN-3376 Add deployment template for bash script
Added a new deployment template - run-bash-script - for
AWS, GCP and Azure. It expects a GitCode artifact and the
name of a Bash script file that is present in the Git repo.
A list of parameters to the script can be provided and are
passed to the script when it is run.

All of the env vars (secret and not) are made available
to the script as well.

The pipeline exports an env var - CODEPIPES_DEPLOY_ACTION -
which can be used by the called script to distinguish between
"deploy" and "undeploy".