VAN-3684 Add release to staging pipeline for Azure
Added stages to the existing Staging pipeline to deploy
to GCP staging first and then to Azure staging.

Unfortunately we can't just use the pipeline template that is
used for GCP as the connection to the database is handled completely
differently.

Also updated the existing PR pipeline to use 'ubuntu-latest' instead
of 'ubuntu-18.04' as the latter is being discontinued.