VAN-3734 Add SonarCloud and go-unittest modules
Added sonarcloud.yaml and go-unittest.yaml for app integration
pipelines.

If the unittest modules is in the pipeline before the sonarcloud
module, then the test coverage information will be passed to
SonarCloud.

Also for GCP, switched the go-build module from using the cloud-
builders go image to the standard golang image. The support for
the cloud-builders image seems to be dropping as it is slow to
adopt new versions of golang.

As the sha_tag is now being passed in from the services, the places
where it was being exported as an env var have been removed.