VAN-3784 Automate process to push module containers
In order to avoid rate limiting issues with anonymous connections
to certain Docker registries, we will be hosting any images required
by our pipeline modules in ECR or GCR (maybe ACR in the future).

This introduces a shell script that pulls the relevant images, tags
them appropriately and pushes them up to ECR/GCR. The relevant images
are listed in a container-manifest.yaml file that is processed by
the script.

Also added a make target that logs into the various registries and
runs the script.