VAN-4078 Add helm image to build locally
In order to add "envsubst" as a tool that can be used to preprocess
Helm values.yaml it needs to be in the Helm container used in GCP.
To do this, we need to build the Helm container and push to cldcvr in
Docker Hub.

Changed the relevant deployment templates to use the new cpi-helm
container instead of having each module build it.
Added the preprocessing of values.yaml to replace environment
variables in the basic-helm and private-helm templates.